requirements.txt
Jinja2==2.8
Markdown==2.6.2
MarkupSafe==0.23
PyYAML==3.11
argparse==1.2.1
click==5.1
wsgiref==0.1.2
